Moved Json Pointer docs and applied some changes from code review
