+++ src/main/java/docoverride/json/Examples.java
@@ -82,26 +82,4 @@ public class Examples {
     Boolean boolVal = array.getBoolean(2);
   }
 
-  public void example1Pointers() {
-    // Build a pointer from a string
-    JsonPointer pointer1 = JsonPointer.from("/hello/world");
-    // Build a pointer from an URI
-    JsonPointer pointer2 = JsonPointer.fromURI(URI.create("#/hello/world"));
-    // Build a pointer manually
-    JsonPointer pointer3 = JsonPointer.create()
-      .append("hello")
-      .append("world");
-  }
-
-  public void example2Pointers(JsonPointer objectPointer, JsonObject jsonObject, JsonPointer arrayPointer, JsonArray jsonArray) {
-    // Query a JsonObject
-    Object result1 = objectPointer.queryJson(jsonObject);
-    // Query a JsonArray
-    Object result2 = arrayPointer.queryJson(jsonArray);
-    // Write starting from a JsonObject
-    objectPointer.writeJson(jsonObject, "new element");
-    // Write starting from a JsonObject
-    arrayPointer.writeJson(jsonArray, "new element");
-  }
-
 }
